fix: folders and airdrop export
