+++ src/components/Airdrop/index.tsx
@@ -1,0 +1,2 @@
+export { default as Airdrop } from './Airdrop';
+export { default as ComingSoon } from './ComingSoon';
